% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distrikt}
\alias{distrikt}
\title{Distrikt (districts)}
\format{
An SF object of Swedish districts, with three variables:
\code{code}, \code{name} and \code{geometry}.
}
\usage{
distrikt
}
\description{
Data originally from Lantmäteriet:
https://geotorget.lantmateriet.se
}
\keyword{datasets}
