% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kommuner}
\alias{kommuner}
\title{Kommuner (municipality)}
\format{
An SF object of Swedish municipalities, with three variables:
\code{code}, \code{name} and \code{geometry}.
}
\usage{
kommuner
}
\description{
Data originally from Statistics Sweden (SCB):
https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser/
}
\keyword{datasets}
