% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lan}
\alias{lan}
\title{Län (administrative region)}
\format{
An SF object of Swedish "län" regions, with four variables:
\code{code}, \code{name}, \code{group} and \code{geometry}.
}
\usage{
lan
}
\description{
Data originally from Statistics Sweden (SCB):
https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser/
}
\keyword{datasets}
