% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsdelar}
\alias{landsdelar}
\title{Landsdelar (informal region; 'country parts')}
\format{
A data frame of Swedish "landsdelar" regions, with two variables:
  \code{name} and \code{geometry}.
}
\usage{
landsdelar
}
\description{
Data originally from Statistics Sweden (SCB), but reworked for new grouping:
https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser/
}
\keyword{datasets}
