% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landskap}
\alias{landskap}
\title{Landskap (informal region; 'country regions')}
\format{
An SF object of Swedish "landskap" regions, with three variables:
\code{code}, \code{name} and \code{geometry}.
}
\usage{
landskap
}
\description{
Data originally from Lantmäteriet (as CC0), but reworked by:
https://github.com/perliedman/svenska-landskap
}
\keyword{datasets}
