% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lokal_arbetsmarknad}
\alias{lokal_arbetsmarknad}
\title{Lokal arbetsmarknad (administrative region; 'local labor market')}
\format{
An SF object of Swedish "lokal arbetsmarknad" regions, with three variables:
\code{code}, \code{name} and \code{geometry}.
}
\usage{
lokal_arbetsmarknad
}
\description{
Data originally from Statistics Sweden (SCB):
https://www.scb.se/hitta-statistik/regional-statistik-och-kartor/regionala-indelningar/digitala-granser/
}
\keyword{datasets}
