% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts2}
\alias{nuts2}
\title{NUTS2 (Nomenclature des Unités Territoriales Statistiques)}
\format{
An SF object of Swedish NUTS regions, with three variables:
\code{code}, \code{name} and \code{geometry}.
}
\usage{
nuts2
}
\description{
Data originally from Statistics Sweden (SCB):
https://www.scb.se/vara-tjanster/oppna-data/oppna-geodata/tatorter/
}
\keyword{datasets}
