% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts3}
\alias{nuts3}
\title{NUTS3 (Nomenclature des Unités Territoriales Statistiques). NB: same as "län"}
\format{
An SF object of Swedish NUTS regions, with three variables:
\code{code}, \code{name} and \code{geometry}.
}
\usage{
nuts3
}
\description{
Data originally from Statistics Sweden (SCB):
https://www.scb.se/vara-tjanster/oppna-data/oppna-geodata/tatorter/
}
\keyword{datasets}
