% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{socknar}
\alias{socknar}
\title{Socknar (parishes)}
\format{
An SF object of Swedish church parishes, with five variables:
\code{code}, \code{name}, \code{area_id}, \code{main_area} and \code{geometry}.
}
\usage{
socknar
}
\description{
Data originally from Lantmäteriet:
https://geotorget.lantmateriet.se
}
\keyword{datasets}
