% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tatorter}
\alias{tatorter}
\title{Tätorter (cities; 'densely populated locations')}
\format{
A data frame of Swedish "tätorter" ('cities'), with 12 variables:
  \code{id}, \code{uu_id}, \code{code}, \code{name},
  \code{mun_code}, \code{mun_name}, \code{lan_code}, \code{lan_name},
  \code{area}, \code{pop}, \code{year} and \code{geometry}.
}
\usage{
tatorter
}
\description{
Data originally from Statistics Sweden (SCB):
https://www.scb.se/vara-tjanster/oppna-data/oppna-geodata/tatorter/
}
\keyword{datasets}
