% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsdelar}
\alias{landsdelar}
\title{Landsdelar (informal region; 'country parts')}
\format{
An SF object of Swedish "landsdelar" regions, with three variables:
\code{code}, \code{name} and \code{geometry}.
}
\usage{
landsdelar
}
\description{
Data originally from Lantmäteriet (as CC0), but reworked by:
https://github.com/perliedman/svenska-landskap
}
\keyword{datasets}
